Added 64bit support for instruction typing.
git-svn-id: svn+ssh://bono.sdsc.edu/projects/pmac/svnrepos/pmac/trunk/X86ElfInstrumentor@1781 55925d41-9515-0410-9586-c6f35bcc715f

+++ include/Disassembler.h
@@ -80,7 +80,7 @@ private:
     void get_ops(op_func op, uint32_t bytemode, uint32_t sizeflag);
 
     /* legacy functions from GNU libopcodes. at least some of these should eventually be gone */
-    uint32_t readMemory(long long unsigned int, uint8_t*, unsigned int, disassemble_info*);
+    uint32_t readMemory(uint64_t, uint8_t*, unsigned int, disassemble_info*);
     void ckprefix();
     const char* prefix_name(int pref, int sizeflag);
     void dofloat(uint32_t sizeflag);
